+++ src/encoder/get_cnt.h
@@ -5,8 +5,8 @@
 #ifndef GET_CNT_H
 #define GET_CNT_H
 
-void get_cnt(std::ifstream &fin, int cnt_of_symbols[]);
+void get_cnt(std::ifstream &fin, unsigned long long cnt_of_symbols[]);
 
-void get_cnt(std::string &s, int cnt_of_symbols[]);
+void get_cnt(std::string &s, unsigned long long cnt_of_symbols[]);
 
 #endif //GET_CNT_H
